Refactored keyhandlers and reset
